some changes

+++ tasks/ovsyannikov_n_num_mistm_in_two_str/common/include/common.hpp
@@ -8,7 +8,7 @@
 
 namespace ovsyannikov_n_num_mistm_in_two_str {
 using InType = std::pair<std::string, std::string>;  // Входные данные: пара строк
-using OutType = int;                                 // Выходные данные: количество несовпадений
+using OutType = size_t;                              // Выходные данные: количество несовпадений
 using BaseTask = ppc::task::Task<InType, OutType>;
 using TestType = std::tuple<InType, OutType, std::string>;
 }  // namespace ovsyannikov_n_num_mistm_in_two_str
